requirements.txt
numpy
visdom
joblib
librosa
tqdm
matplotlib
scikit-learn
pyyaml
torch
torchvision
torchaudio
tensorboard